% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{gwf_irt}
\alias{gwf_irt}
\title{Factor scores and IRT estimates of latent dimensions using the \link{gwf} data.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4591 rows and 19 columns.
}
\usage{
gwf_irt
}
\description{
The factor scores are replicated from Wright, Joseph. 2021. “The Latent
Characteristics That Structure Autocratic Rule.” Political Science Research
and Methods 9 (1): 1–19. https://doi.org/10.1017/psrm.2019.50. See the
\href{https://xmarquez.github.io/GWFdata/articles/replicating-wright-2019.html}{vignette}
for details.
}
\details{
\describe{

\item{gwf_country}{name of country}

\item{cowcode}{Correlates of War country code. Countries  that have changed
territory during the period from 1946 to 2010 have the following country
codes in this data set: Czechoslovakia (315); East Germany (265); Ethiopia
(530); Eritrea (531); Russia and Soviet Union (365); South Vietnam (817);
South Yemen (680); Vietnam/North Vietnam (816); Yemen/ North Yemen (678);
Yugoslavia/Serbia (435). Note that regimes that either lost (Ethiopia) or
gained territory (Vietnam/North Vietnam) during the lifetime of the regime
have the same country code throughout the entire regime spell. Users should
note this feature of the data set when merging with other data sets.}

\item{year}{calendar year}

\item{pers_irt_11_F1}{IRT scores for personalism, 11 items}

\item{pers_irt_11_F1}{SE for IRT scores for personalism, 11 items}

\item{pers_irt_10_F1}{IRT scores for personalism, 10 items}

\item{pers_irt_10_F1}{SE for IRT scores for personalism, 10 items}

\item{pers_irt_8_F1}{IRT scores for personalism, 8 items}

\item{pers_irt_8_F1}{SE for IRT scores for personalism, 8 items}

\item{pers_irt_11_full_info_F1}{IRT scores for personalism, full information, 11 items}

\item{pers_irt_11_full_info_F1}{SE for IRT scores for personalism, full information, 11 items}

\item{party_irt_11_F1}{IRT scores for party, 11 items}

\item{party_irt_11_F1}{SE for IRT scores for party, 11 items}

\item{military_irt_11_F1}{IRT scores for party, 11 items}

\item{military_irt_11_F1}{SE for IRT scores for party, 11 items}

\item{party_factor}{Scores on first factor of factor analysis (party).}

\item{military_factor}{Scores on second factor of factor analysis (military).}

\item{personalism_factor}{Scores on third factor of factor analysis (personalism).}

\item{fourth_factor}{Scores on fourth factor of factor analysis (other).}

}
}
\keyword{datasets}
