% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWFdata-package.R
\docType{package}
\name{GWFdata-package}
\alias{GWFdata}
\alias{GWFdata-package}
\title{GWFdata: Replication data for the book "How Dictatorships Work" by Barbara Geddes, Joseph Wright, and Erica Frantz}
\description{
Contains the replication dataset for book "How Dictatorships Work" by Barbara Geddes, Joseph Wright, and Erica Frantz, including the codebook.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/xmarquez/GWFdata}
  \item Report bugs at \url{https://github.com/xmarquez/GWFdata/issues}
}

}
\author{
\strong{Maintainer}: Xavier Marquez \email{marquez.x@gmail.com}

}
\keyword{internal}
